pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asciicker-pipeline-v2"
version = "0.1.0"
description = "Greenfield sprite-sheet to XP pipeline MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Flask>=3.0",
  "Pillow>=10.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
